Adding LSA Messages and LSA sending packets

+++ src/WeightedGraph.java
@@ -2,11 +2,11 @@
  * the weighted graph class for dijkstar algorithm
  */
 public class WeightedGraph {
-    private int [][]  edges;  // adjacency matrix
+    private double [][]  edges;  // adjacency matrix
     private Object [] labels;
 
     public WeightedGraph (int n) {
-        edges  = new int [n][n];
+        edges  = new double [n][n];
         labels = new Object[n];
     }
 
@@ -15,10 +15,10 @@ public class WeightedGraph {
     public void   setLabel (int vertex, Object label) { labels[vertex]=label; }
     public Object getLabel (int vertex)               { return labels[vertex]; }
 
-    public void    addEdge    (int source, int target, int w)  { edges[source][target] = w; }
+    public void    addEdge    (int source, int target, double w)  { edges[source][target] = w; }
     public boolean isEdge     (int source, int target)  { return edges[source][target]>0; }
     public void    removeEdge (int source, int target)  { edges[source][target] = 0; }
-    public int     getWeight  (int source, int target)  { return edges[source][target]; }
+    public double     getWeight  (int source, int target)  { return edges[source][target]; }
 
     public int [] neighbors (int vertex) {
         int count = 0;
